CS: fix regex layout
Using the `x` modifier allows for using comments within a regex itself removing the need for the concatenation previously used.
The `x` modifier ignores all whitespace within the regex, allowing for a nice layout as well.
To facilitate this, I've moved the regex delimiters up into the regex string, so the modifier and the regex stay together.

N.B.: There are already unit tests in place which cover this method comprehensively, so if I did things right, those should still pass.

Refs:
* http://php.net/manual/en/reference.pcre.pattern.modifiers.php