CS: minor regex efficiency improvement
No need to remember sub-matches which are not being used anyway (previous match 2 and 3). The parentheses were in this case also not needed for grouping, so removing them was the simplest fix.

N.B.: There are already unit tests in place which cover this method comprehensively, so if I did things right, those should still pass.