I18n: Make the text slightly more plugin agnostic
AFAIK WHIP is intended to also be usable by other (non-Yoast) plugins/themes.
There were currently two phrases which made this awkward.

This PR changes those.

For the first phrase, I've chosen to change `WordPress and Yoast SEO` to `WordPress and all it's plugins and themes`. This incidentally fixes issue 35.
If it is preferred that `Yoast SEO` not be removed, I would suggest the following:
* Move the name out of the text string and put it back in using `sprintf()`.
* Make the name filterable to allow other plugins to pass in their plugin name.
    - Potentially, it could even be turned into an array which could be passed into the string using `implode()` to allow for a number of plugins to be mentioned.

For the second phrase, just changing the `our` in `vetted by our Yoast support team` to `the` improves the re-usability.

Fixes 35