QA: use the most specific PHPUnit assertion possible
This is a long established best practice.

PHPUnit contains a variety of assertions and the ones available has been extended  hugely over the years.
To have the most reliable tests, the most specific assertion should be used.

This implements this for the WHIP plugin, while keeping in mind that at this time, PHPUnit 3.x should still be supported.

Refs:
* https://phpunit.de/manual/4.8/en/appendixes.assertions.html
* https://phpunit.readthedocs.io/en/7.5/assertions.html#

Most notably, this changes calls to `assertEquals()` to `assertSame()`, where `assertEquals()` does a loose type comparison and `assertSame()` does a strict type comparison.

Refs:
* https://phpunit.de/manual/6.5/en/appendixes.assertions.html#appendixes.assertions.assertEquals
* https://phpunit.de/manual/6.5/en/appendixes.assertions.html#appendixes.assertions.assertSame