Travis: fix PHP 5.2 build
Composer was running out of memory.

Doing the `composer remove`s without installing, should lower the memory usage as the `composer install` which is run after the removals anyway will now only need to install what we really need for the builds.

Includes removing the reference to the Dealerdirect plugin from the Travis script. This has not been a direct dependency since 74, it is now a dependency of YoastCS.