Tests: add base TestCase
To make the tests cross-version compatible for the WHIP module, not that much is needed in terms of compatibility polyfills.

The most pertinent issue is to do with testing exceptions.

This commit adds a base `TestCase` from which all WHIP test classes will extend from now on. This `TestCase` class only contains a helper method to test exceptions in a PHPUnit cross-version compatible manner.

Includes adding the new testCase to the PHPCS ruleset so sniffs which behave differently in test situations will recognize these correctly.