Tests: autoload test files via Composer
The Composer `autoload-dev` directive allows for specifying files which should only be autoloadable in a `dev` environment.

By adding this directive to the `composer.json` file, we can get rid of most `require_once` statements in the test suite.

The only one remaining is the loading of the WP Core functions mocks. This could be done via a `files` directive within `autoload-dev`, but then they would always be loaded, while they are only needed in the context of running the tests.

So instead, I've added the `require` for the WP Core functions mock file to the test `bootstrap` file.

A side-effect of this change is that we can get rid of the `setUpBeforeClass()` methods which are problematic for PHPUnit cross-version compatibility due to the addition of the `void` return type to the method signature in PHPUnit 8.