Composer: update PHP Parallel Lint
Update version constraint for PHP Parallel Lint after new release.

The most significant change in this release is that PHP Parallel Lint now also supports PHP 5.3. This allows us to make some simplifications to the Travis script as we, for the most part, no longer need to special case PHP 5.3.

Refs:
* https://github.com/php-parallel-lint/PHP-Parallel-Lint/releases/tag/v1.3.0