Whip_WPMessageDismissListener::listen(): bug fix
The `Whip_MessageDismisser::verifyNonce()` method expects to be passed a `string`, but if the nonce was not available, it would be passed `null`, which could lead to PHP 8.1 "passing null to non-nullable" deprecation notices (though in this case wouldn't as the underlying [`wp_verify_nonce()`](https://developer.wordpress.org/reference/functions/wp_verify_nonce/) function casts the value to string anyway).

All the same, it is a bug. Fixed now.

Note: as the test suite does not test with a real WP setup, this bug currently cannot easily be safeguarded via a test.

The `WHIP_WPMessageDismissListener::listen()` method does have tests, but mocks away the `Whip_MessageDismisser::verifyNonce()` method, so the mock was hiding the bug.