Tests: minor modernization
As of PHPUnit 4.8.36 and PHPUnit 5.7.21, PHPUnit contains a _forward_ compatibility layer for the most commonly used PHPUnit classes to make handling the change to namespaced classes in PHPUnit 6.0 more straight-forward.

As support for PHP 5.2 has been dropped from this plugin, we can now use this forward-compatibility layer and no longer need to create a class alias for the PHPUnit `TestCase` ourselves.